Compute CalcCharaSelect window x in double precision
The inlined MC-window centering computes win->x as
(double)(float)(0x280 - winWidth) * 0.5 in double (matching Ghidra and
the win->y term just below), not with a single-precision multiply.
CalcCharaSelect 44.56% -> 45.20%.

+++ src/wm_menu.cpp
@@ -7524,8 +7524,8 @@ void CMenuPcs::CalcCharaSelect()
 			winWidth = static_cast<short>((winWidth + 2) * 0x16 + 0x40);
 			const short winHeight = static_cast<short>(*winMess) * 0x1E + 0x40;
 			MenuWindowInfo* const win = m_menuWindowInfo;
-			win->x = static_cast<short>(static_cast<int>(static_cast<float>(0x280 - winWidth) *
-			                                             static_cast<float>(DOUBLE_803313f8)));
+			win->x = static_cast<short>(static_cast<int>(static_cast<double>(static_cast<float>(0x280 - winWidth)) *
+			                                             DOUBLE_803313f8));
 			win->y = static_cast<short>(static_cast<int>((static_cast<double>(FLOAT_80331430 -
 			                                                                  static_cast<float>(winHeight))) *
 			                                             DOUBLE_803313f8));
